requirements.txt
google-genai
pillow
numpy
openai>=1.0.0
matplotlib
python-dotenv
